nodehw5

+++ src/controllers/contacts.js
@@ -11,7 +11,13 @@ import { parsePaginationParams } from '../utils/parsePaginationParams.js';
 import { parseSortParams } from '../utils/parseSortParams.js';
 import { parseFilterParams } from '../utils/parseFilterParams.js';
 
-export const getContactsController = async (req, res) => {
+export const getContactsController = async (req, res, next) => {
+  const userId = req.user._id;
+  if (!userId) {
+    next(createHttpError(403, 'You do not have permission to access this contact!'));
+    return;
+  }
+  
   const { page, perPage } = parsePaginationParams(req.query);
   const { sortBy, sortOrder } = parseSortParams(req.query);
   const filter = parseFilterParams(req.query)
@@ -33,6 +39,12 @@ export const getContactsController = async (req, res) => {
 
 export const getContactByIdController = async (req, res, next) => {
   const { contactId } = req.params;
+  const userId = req.user._id;
+  if (!userId) {
+    next(createHttpError(403, 'You do not have permission to access this contact!'));
+    return;
+  }
+
   if (!mongoose.Types.ObjectId.isValid(contactId)) {
     return res.status(400).json({
       status: 400,
@@ -55,7 +67,7 @@ export const getContactByIdController = async (req, res, next) => {
 };
 
 export const createContactController = async (req, res) => {
-  const contact = await createContact(req.body);
+  const contact = await createContact({ ...req.body, userId: req.user._id });
 
   res.status(201).json({
     status: 201,
@@ -66,6 +78,12 @@ export const createContactController = async (req, res) => {
 
 export const patchContactController = async (req, res, next) => {
   const { contactId } = req.params;
+  const userId = req.user._id;
+  if (!userId) {
+    next(createHttpError(403, 'You do not have permission to access this contact!'));
+    return;
+  }
+
   const result = await updateContact(contactId, req.body);
   if (!result) {
     next(createHttpError(404, 'Contact not found'));
@@ -80,6 +98,12 @@ export const patchContactController = async (req, res, next) => {
 
 export const deleteContactController = async (req, res, next) => {
   const { contactId } = req.params;
+  const userId = req.user._id;
+  if (!userId) {
+    next(createHttpError(403, 'You do not have permission to access this contact!'));
+    return;
+  }
+
   const contact = await deleteContact(contactId);
   if (!contact) {
     next(createHttpError(404, 'Contact not found'));
